Extract streaming worker helpers
